Split functions configuration from FeaturesConfig
