+++ presto-main/src/main/java/com/facebook/presto/metadata/FunctionAndTypeManager.java
@@ -52,6 +52,7 @@ import com.facebook.presto.spi.function.SqlFunctionId;
 import com.facebook.presto.spi.function.SqlInvokedFunction;
 import com.facebook.presto.sql.analyzer.FeaturesConfig;
 import com.facebook.presto.sql.analyzer.FunctionAndTypeResolver;
+import com.facebook.presto.sql.analyzer.FunctionsConfig;
 import com.facebook.presto.sql.analyzer.TypeSignatureProvider;
 import com.facebook.presto.sql.gen.CacheStatsMBean;
 import com.facebook.presto.sql.tree.QualifiedName;
@@ -134,12 +135,13 @@ public class FunctionAndTypeManager
             TransactionManager transactionManager,
             BlockEncodingSerde blockEncodingSerde,
             FeaturesConfig featuresConfig,
+            FunctionsConfig functionsConfig,
             HandleResolver handleResolver,
             Set<Type> types)
     {
         this.transactionManager = requireNonNull(transactionManager, "transactionManager is null");
         this.blockEncodingSerde = requireNonNull(blockEncodingSerde, "blockEncodingSerde is null");
-        this.builtInTypeAndFunctionNamespaceManager = new BuiltInTypeAndFunctionNamespaceManager(blockEncodingSerde, featuresConfig, types, this);
+        this.builtInTypeAndFunctionNamespaceManager = new BuiltInTypeAndFunctionNamespaceManager(blockEncodingSerde, functionsConfig, types, this);
         this.functionNamespaceManagers.put(DEFAULT_NAMESPACE.getCatalogName(), builtInTypeAndFunctionNamespaceManager);
         this.functionInvokerProvider = new FunctionInvokerProvider(this);
         this.handleResolver = requireNonNull(handleResolver, "handleResolver is null");
@@ -152,13 +154,19 @@ public class FunctionAndTypeManager
                 .build(CacheLoader.from(key -> resolveBuiltInFunction(key.functionName, fromTypeSignatures(key.parameterTypes))));
         this.cacheStatsMBean = new CacheStatsMBean(functionCache);
         this.functionSignatureMatcher = new FunctionSignatureMatcher(this);
-        this.typeCoercer = new TypeCoercer(featuresConfig, this);
+        this.typeCoercer = new TypeCoercer(functionsConfig, this);
         this.nativeExecution = featuresConfig.isNativeExecutionEnabled();
     }
 
     public static FunctionAndTypeManager createTestFunctionAndTypeManager()
     {
-        return new FunctionAndTypeManager(createTestTransactionManager(), new BlockEncodingManager(), new FeaturesConfig(), new HandleResolver(), ImmutableSet.of());
+        return new FunctionAndTypeManager(
+                createTestTransactionManager(),
+                new BlockEncodingManager(),
+                new FeaturesConfig(),
+                new FunctionsConfig(),
+                new HandleResolver(),
+                ImmutableSet.of());
     }
 
     public FunctionAndTypeResolver getFunctionAndTypeResolver()
