+++ presto-main/src/test/java/com/facebook/presto/sql/planner/TestExpressionDomainTranslator.java
@@ -20,7 +20,7 @@ import com.facebook.presto.common.predicate.ValueSet;
 import com.facebook.presto.common.type.DecimalType;
 import com.facebook.presto.common.type.Type;
 import com.facebook.presto.metadata.Metadata;
-import com.facebook.presto.sql.analyzer.FeaturesConfig;
+import com.facebook.presto.sql.analyzer.FunctionsConfig;
 import com.facebook.presto.sql.planner.ExpressionDomainTranslator.ExtractionResult;
 import com.facebook.presto.sql.tree.BetweenPredicate;
 import com.facebook.presto.sql.tree.Cast;
@@ -1168,7 +1168,7 @@ public class TestExpressionDomainTranslator
     @Test
     public void testLegacyCharComparedToVarcharExpression()
     {
-        metadata = createTestMetadataManager(new FeaturesConfig().setLegacyCharToVarcharCoercion(true));
+        metadata = createTestMetadataManager(new FunctionsConfig().setLegacyCharToVarcharCoercion(true));
         literalEncoder = new LiteralEncoder(metadata.getBlockEncodingSerde());
         domainTranslator = new ExpressionDomainTranslator(literalEncoder);
