+++ presto-tests/src/test/java/com/facebook/presto/tests/TestQuerySpillLimits.java
@@ -16,6 +16,7 @@ package com.facebook.presto.tests;
 import com.facebook.presto.Session;
 import com.facebook.presto.spiller.NodeSpillConfig;
 import com.facebook.presto.sql.analyzer.FeaturesConfig;
+import com.facebook.presto.sql.analyzer.FunctionsConfig;
 import com.facebook.presto.testing.LocalQueryRunner;
 import com.facebook.presto.testing.QueryRunner;
 import com.facebook.presto.tpch.TpchConnectorFactory;
@@ -78,6 +79,7 @@ public class TestQuerySpillLimits
                 new FeaturesConfig()
                         .setSpillerSpillPaths(spillPath.getAbsolutePath())
                         .setSpillEnabled(true),
+                new FunctionsConfig(),
                 nodeSpillConfig,
                 false,
                 true);
